use the api as a macro
